Add AddSchemeDialog skeleton + Restructur files
